+++ vite.config.ts
@@ -2,6 +2,7 @@ import { defineConfig } from "vite";
 
 export default defineConfig({
   build: {
+    target: "es2019",
     outDir: "dist",
     sourcemap: true,
     rollupOptions: {